+++ CHANGELOG.md
@@ -5,11 +5,13 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ## [Unreleased]
 ### Added
+- Added [`no-class`] rule
 - Added [`no-let`] rule
 - Added [`no-mutating-assign`] rule
 - Added [`no-mutation`] rule
 - Added [`no-this`] rule
 
+[`no-class`]: ./docs/rules/no-class.md
 [`no-let`]: ./docs/rules/no-let.md
 [`no-mutating-assign`]: ./docs/rules/no-mutating-assign.md
 [`no-mutation`]: ./docs/rules/no-mutation.md